removes otel_span_processor behavior attribute from baggage_processor
specifying the behavior produces a warning unless the sdk is included,
but including it was causing issues in one particular application
(started via a mix task which may or may not be related). removing the
attribute doesn't seem to cause any problems, or produce any warnings,
but removing the sdk dep _does_ fix the issue in that one application.